+++ clubs.json
@@ -3,7 +3,7 @@
         {
             "name": "Simply Lift",
             "email": "john@simplylift.co",
-            "points": "19",
+            "points": "18",
             "bookings": [
                 {
                     "competition": "Spring Festival",
