+++ competitions.json
@@ -3,12 +3,12 @@
         {
             "name": "Spring Festival",
             "date": "2020-03-27 10:00:00",
-            "numberOfPlaces": "17"
+            "numberOfPlaces": "1"
         },
         {
             "name": "Fall Classic",
             "date": "2020-10-22 13:30:00",
-            "numberOfPlaces": "17"
+            "numberOfPlaces": "6"
         }
     ]
 }
