--- tests/test_home.py
@@ -1,5 +1,0 @@
-def test_index(test_client):
-    """Test the index route."""
-    response = test_client.get("/")
-    assert response.status_code == 200
-    assert b"Welcome to the GUDLFT Registration Portal!" in response.data
